fixed problem with some non 7bit ASCII path/file names. Bug #3144278
git-svn-id: http://svn.code.sf.net/p/megui/code@1879 a21b8521-4df1-4ed3-b48c-cff5bc73dcb3

+++ megui/trunk/core/details/CleanupJob.cs
@@ -91,6 +91,8 @@ namespace MeGUI.core.details
             {
                 try
                 {
+                    if (!File.Exists(file))
+                        continue;
                     File.Delete(file);
                     i.LogEvent("Successfully deleted " + file);
                 }
